Preparation work, user login working, and cleanup
Most of the changes are cleaning up the project and setting up better organization.

I got rid of geodjango, since it is a bit overkill for this and basically requires postgres. Plus, sqlite is much easier to develop on.

I will upload the sass code after removing code I copied from elsewhere.

Next up is moving the websockets server into django, currently it is a node.js server(which is not on git).